Fix database integration crash and save layout after folder

+++ app/src/main/java/foundation/e/blisslauncher/core/database/model/LauncherItem.java
@@ -67,7 +67,7 @@ public class LauncherItem {
      * {@link Constants#CONTAINER_HOTSEAT})
      */
     @ColumnInfo(name = "screen_id")
-    public long screenId = -1;
+    public int screenId = -1;
 
     /**
      * Indicates the position of the associated cell.
